+++ CHANGELOG.md
@@ -10,3 +10,8 @@ cruhl - Clean up and making column widths variable based on content.
 -----
 cruhl - Re-worked printf calculation to use colors rather than "-f" for frozen.
 	Limited checking for frozen to "Latest" and the primary environmnet
+
+2.1
+-----
+cruhl - Allow multiple primaries
+	Show obsolete constraints